Optimized version working correctly. Still slow.

+++ wordcloud/wordcloud.h
@@ -9,5 +9,5 @@ std::vector<short> countVec(1400000000);
 std::vector<std::string> wordcloud_vec(bloomBits);
 std::vector<std::string> badWords = {"value not found"};
 std::set<std::string> uniqWords;
-std::vector<short> word_list_vec;
+std::vector<size_t> word_list_vec;
